+++ TravelList/AddTravelViewController.swift
@@ -38,7 +38,7 @@ class AddTravelViewController: UIViewController {
         if sender as? NSObject != self.doneButton{
             return
         }
-        if let itemName = self.textfield.text?.stringByTrimmingCharactersInSet(NSCharacterSet.whitespaceAndNewlineCharacterSet()) where !itemName.isEmpty{
+        if let itemName = TravelListItem.normalizedName(self.textfield.text) {
             self.travelItem = TravelListItem(name: itemName)
         }
     }